+++ page.php
@@ -38,6 +38,6 @@
 	</section>
 	<!-- /Section -->
 	
-<?php get_sidebar(); ?>
+<?php //get_sidebar(); ?>
 
 <?php get_footer(); ?>